pratica10/my_project/requirements.txt
Django>=5.2
Faker
djangorestframework
markdown
django-filter